backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
pandas>=2.0
python-multipart>=0.0.6
reportlab>=4.0   # for simple PDF generation
django-cors-headers>=3.15
